Fix warning about coveralls not needing quotes.
When using ueberauth_auth0 in a project, the below warning is emmitted.
This PR clears the warning.

warning: found quoted keyword "coveralls" but the quotes are not required. Note that keywords are always atoms, even when quoted. Similar to atoms, keywords made exclusively of Unicode letters, numbers, underscore, and @ do not require quotes
  project_root/deps/ueberauth_auth0/mix.exs:25