Handle 403 coming from Auth0
In case a user enters their own code and hits the callback URL, the
request will always raise OAuth2.Error. This is not ideal, as it means
a 500 being shown to the user.

What should happen instead, is that we populated the ueberauth_failure
assigns and the implementer can decide what to show to the the user.